refactor(harness,graph): convert no_progress.rs and langfuse.rs to module dirs
Both flat files violated the dir/{mod.rs,types.rs,test.rs} convention.
- harness/no_progress.rs -> no_progress/{mod.rs,types.rs,test.rs}
  (ToolAttempt/NoProgress/LadderState/NoProgressTracker type defs
  split from the escalation-ladder impl).
- harness/observability/langfuse.rs -> langfuse/{mod.rs,types.rs,test.rs}
  (LangfuseAuth/LangfuseTraceConfig/LangfuseClient type defs split
  from the client impl and payload helpers).
- graph/observability/langfuse/tests.rs renamed to test.rs to unify
  naming with the harness module and the repo-wide test.rs convention.

Public API paths (re-exported from harness::no_progress and
harness::observability) unchanged. No behavior change.
