+++ src/graph/observability/langfuse.rs
@@ -484,4 +484,4 @@ fn pop<K: std::hash::Hash + Eq>(
 }
 
 #[cfg(test)]
-mod tests;
+mod test;
